- Added page with list of coins - Added a page with a detailed description of the coin - Added WpfPlot library + 2 price charts implemented(list & detailed pages) - 3 public WebApi methods for CoinGecko implemented

+++ client/desktop/CryptoPro.WpfApp/MVVM/Models/Sparkline.cs
@@ -2,5 +2,5 @@
 
 public class Sparkline
 {
-    public List<decimal> Prices { get; set; }
+    public List<double> Prices { get; set; }
 }
